+++ sass/labo-components.scss
@@ -11,6 +11,7 @@
 @import 'annotation/index.scss';
 @import 'collection/index.scss';
 @import 'flex/index.scss';
+@import 'helpers/index.scss';
 @import 'linkeddata/index.scss';
 @import 'player/index.scss';
 @import 'project/index.scss';
@@ -24,5 +25,3 @@
 .bg__no-bootstrap-custom-styling {
   @import './workspace/index.scss';
 }
-
-@import "util/index";
